backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# AI & ML
openai-clip==1.0.1
torch==2.1.2
torchvision==0.16.2
transformers==4.36.2
sentence-transformers==2.3.1
google-generativeai==0.3.2
langchain==0.1.4
langchain-google-genai==0.0.6
langgraph==0.0.20
langchain-core==0.1.16

# Vector DB & Cache
qdrant-client==1.7.3
redis==5.0.1

# Data Processing
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.4.0
chonkie==0.0.3

# RAG & Evaluation
ragas==0.1.1
datasets==2.16.1

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
pillow==10.2.0
httpx==0.26.0
aiofiles==23.2.1
